third_party/blink/web_tests/platform/generic/editing/pasteboard/avoid-copying-body-with-background-expected.txt

This test copies and pastes content inside a body with background and text-decoration properties. WebKit should not nest body.
| "
"
| <ul>
|   "
"
|   <li>
|     "hello, world "
|     <a>
|       href="http://www.webkit.org/"
|       "WebKit"
|   "
"
|   <li>
|     "Copied content"
|   <li>
|     "hello, world "
|     <a>
|       href="http://www.webkit.org/"
|       "WebKit"
|   <li>
|     "Copied content<#selection-caret>"
|   "
"
| "
"
| "
"
